+++ CHANGELOG.md
@@ -54,9 +54,10 @@ The major version bump is due to upping the required PHP version from `8.5` to `
 	- `ImageInterface::toDataUri()`
 	- `ImageInterface::toStream()`
 	- `ImageInterface::getMimeType()`
+* Added `Scale` operation to the pixel library.
 * Added operation `Pipeline` class to the pixel library.
 * It is now possible to implement custom image inspectors to retrieve information from the images.
-* Added a `TopColors` inspector to replace the `ImageInterface::getTopColors()` method.
+* Added `TopColors` inspector to the pixel library.
 * Added dotenv loader.
 * Added `Signer::validateOrThrow()` method.
 * Added support for all new Redis commands added in Redis 8.2, 8.4, 8.6, 8.8 and 8.10 to the Redis client.